+++ Sources/Extensions/Foundation/URLExtensions.swift
@@ -55,5 +55,36 @@ public extension URL {
 	public mutating func appendQueryParameters(_ parameters: [String: String]) {
 		self = appendingQueryParameters(parameters)
 	}
-
 }
+
+#if os(iOS) || os(tvOS)
+import UIKit
+import AVFoundation
+    
+// MARK: - Methods
+public extension URL {
+    
+    /// Generate a thumbnail image from given url. Returns nil if no thumbnail could be created. This function may take some time to complete. It's recommended to dispatch the call if the thumbnail is not generated from a local resource.
+    ///
+    ///     var url = URL(string: "https://video.golem.de/files/1/1/20637/wrkw0718-sd.mp4")!
+    ///     var thumbnail = url.thumbnail()
+    ///     thumbnail = url.thumbnail(fromTime: 5)
+    ///
+    ///     DisptachQueue.main.async {
+    ///         someImageView.image = url.thumbnail()
+    ///     }
+    ///
+    /// - Parameter time: Seconds into the video where the image should be generated.
+    /// - Returns: The UIImage result of the AVAssetImageGenerator
+    public func thumbnail(fromTime time: Float64 = 0) -> UIImage? {
+        let imageGenerator = AVAssetImageGenerator(asset: AVAsset(url: self))
+        let time = CMTimeMakeWithSeconds(time, 1)
+        var actualTime = CMTimeMake(0, 0)
+        
+        guard let cgImage = try? imageGenerator.copyCGImage(at: time, actualTime: &actualTime) else {
+            return nil
+        }
+        return UIImage(cgImage: cgImage)
+    }
+}
+#endif